[ORGA] add tests for deleteParticipationModal component
